+++ EspionSpotify/Updater.cs
@@ -1,8 +1,8 @@
-﻿using EspionSpotify.Properties;
-using System;
+﻿using System;
 using System.Configuration;
 using System.IO;
 using System.Windows.Forms;
+using EspionSpotify.Properties;
 
 namespace EspionSpotify
 {
@@ -13,7 +13,10 @@ namespace EspionSpotify
         private static readonly string ProjectDirectory = AppDomain.CurrentDomain.BaseDirectory;
 
         private static readonly string UpdaterExtractedTempDirectoryPath = $"{ProjectDirectory}{UPDATER_TMP_DIRECTORY}";
-        private static readonly string UpdaterExtractedContentDirectoryPath = $@"{UpdaterExtractedTempDirectoryPath}\{UPDATER_DIRECTORY}";
+
+        private static readonly string UpdaterExtractedContentDirectoryPath =
+            $@"{UpdaterExtractedTempDirectoryPath}\{UPDATER_DIRECTORY}";
+
         private static readonly string UpdaterDirectoryPath = $"{ProjectDirectory}{UPDATER_DIRECTORY}";
 
         internal static void UpgradeSettings()
@@ -24,11 +27,12 @@ namespace EspionSpotify
             UpdateDirectoryContent();
             DeleteOlderUserSettings();
         }
+
         private static void UpdateDirectoryContent()
         {
             try
             {
-                Directory.Delete(UpdaterDirectoryPath, recursive: true);
+                Directory.Delete(UpdaterDirectoryPath, true);
                 Directory.Move(UpdaterExtractedContentDirectoryPath, UpdaterDirectoryPath);
                 Directory.Delete(UpdaterExtractedTempDirectoryPath);
             }
@@ -40,15 +44,14 @@ namespace EspionSpotify
 
         private static void DeleteOlderUserSettings()
         {
-            var path = Path.GetFullPath(Path.Combine(ConfigurationManager.OpenExeConfiguration(ConfigurationUserLevel.PerUserRoamingAndLocal).FilePath, @"..\..\"));
+            var path = Path.GetFullPath(Path.Combine(
+                ConfigurationManager.OpenExeConfiguration(ConfigurationUserLevel.PerUserRoamingAndLocal).FilePath,
+                @"..\..\"));
             var settingPaths = Directory.GetDirectories(path);
             foreach (var settingPath in settingPaths)
             {
-                DirectoryInfo info = new DirectoryInfo(settingPath);
-                if (info.Name != Application.ProductVersion)
-                {
-                    Directory.Delete(settingPath, true);
-                }
+                var info = new DirectoryInfo(settingPath);
+                if (info.Name != Application.ProductVersion) Directory.Delete(settingPath, true);
             }
         }
     }
